Fix llvm/clang major requires
If an application or library compiles with rocm-device-libs, the major
version of clang used must match the version used to compile
rocm-devices-libs.

To enforce this, this package needs to explicitly require the correct clang
package. Furthermore, while this package doesn't strictly require clang, the
bitcode isn't really useful without it.