Change to use forked sources
Since upstream does not target stable LLVM, it's a real hassle to
maintain this package as-is. Fortunately, forking the source and
backporting patches is pretty trivial.

I setup a new fork:
https://github.com/Mystro256/ROCm-Device-Libs/

and I will be tagging using the LLVM major number + a patch number as I
backport changes. E.g. 16.0 is the first tag that targets LLVM 16, while
16.1 will have bugfix backports from newer ROCm releases.

If you update LLVM, just change the bugfix_version to 0, as I prepare
the intial *.0 tag as soon after the upstream LLVM release branch is
created. E.g. I created 16.0 about a month ago.